Small ESP32 code change
"poll()" methods for TDS and conductivity sensor no longer need to pass in the temperature sensor as an argument as it's now passed in the constructor.
Changed the mcp object from a const reference to a const ptr for every sensor in order to avoid copying the mcp object for each of them.